dividing the big files
