+++ flatters/flatten_metadata.go
@@ -1,0 +1,77 @@
+package flatters
+
+import (
+	"fmt"
+	"strings"
+
+	v1alpha1 "github.com/gardener/gardener/pkg/apis/core/v1alpha1"
+	"github.com/hashicorp/terraform/helper/schema"
+	v1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"github.com/kyma-incubator/terraform-provider-gardener/expand"
+)
+
+func FlattenMetadata(meta metav1.ObjectMeta, d *schema.ResourceData, metaPrefix ...string) []interface{} {
+	m := make(map[string]interface{})
+	prefix := ""
+	if len(metaPrefix) > 0 {
+		prefix = metaPrefix[0]
+	}
+	configAnnotations := d.Get(prefix + "metadata.0.annotations").(map[string]interface{})
+	m["annotations"] = expand.RemoveInternalKeys(meta.Annotations, configAnnotations)
+	if meta.GenerateName != "" {
+		m["generate_name"] = meta.GenerateName
+	}
+	configLabels := d.Get(prefix + "metadata.0.labels").(map[string]interface{})
+	m["labels"] = expand.RemoveInternalKeys(meta.Labels, configLabels)
+	m["name"] = meta.Name
+	m["resource_version"] = meta.ResourceVersion
+	m["self_link"] = meta.SelfLink
+	m["uid"] = fmt.Sprintf("%v", meta.UID)
+	m["generation"] = meta.Generation
+
+	if meta.Namespace != "" {
+		m["namespace"] = meta.Namespace
+	}
+
+	return []interface{}{m}
+}
+
+func flattenLocalObjectReference(in *v1.LocalObjectReference) []interface{} {
+	att := make(map[string]interface{})
+	if in.Name != "" {
+		att["name"] = in.Name
+	}
+	return []interface{}{att}
+}
+
+
+func IdParts(id string) (string, string, error) {
+	parts := strings.Split(id, "/")
+	if len(parts) != 2 {
+		err := fmt.Errorf("Unexpected ID format (%q), expected %q", id, "namespace/name")
+		return "", "", err
+	}
+
+	return parts[0], parts[1], nil
+}
+
+func BuildID(meta metav1.ObjectMeta) string {
+	return meta.Namespace + "/" + meta.Name
+}
+
+func newStringSet(f schema.SchemaSetFunc, in []string) *schema.Set {
+	var out = make([]interface{}, len(in), len(in))
+	for i, v := range in {
+		out[i] = v
+	}
+	return schema.NewSet(f, out)
+}
+
+func newCIDRSet(f schema.SchemaSetFunc, in []v1alpha1.CIDR) *schema.Set {
+	var out = make([]interface{}, len(in), len(in))
+	for i, v := range in {
+		out[i] = string(v)
+	}
+	return schema.NewSet(f, out)
+}
