+++ validator/validator.go
@@ -1,4 +1,4 @@
-package helper
+package validator
 
 import (
 	"fmt"
